pyproject.toml
[tool.poetry]
name = "django-catalog"
version = "0.1.0"
description = ""
authors = ["5alex52 <alkurkovsk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-cleanup = "^9.0.0"
django-debug-toolbar = "^4.4.6"
pillow = "^11.0.0"
pylint = "^3.3.2"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
django-material-admin = "^1.8.6"
pytz = "^2024.2"
requests = "^2.32.3"
pydantic-settings = "^2.6.1"
gunicorn = "^23.0.0"
black = "^24.10.0"
reorder-python-imports = "^3.14.0"
easy-thumbnails = "^2.10"
pytils = "^0.4.1"
django-silk = "^5.3.1"
drf-spectacular = "^0.28.0"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.51"
telebot = "^0.0.5"
pre-commit = "^4.0.1"
django-unfold = "^0.42.0"
django-smart-selects = "^1.7.2"
django-autocomplete-light = "^3.11.0"
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "goodtoeat.settings"
python_files = ["tests.py", "test_*.py", "*_test.py", "tests/*/*_tests.py"]
addopts = "-p no:warnings"

[tool.pytest_env]
ENVIROMENT="testing"

[tool.black]
line-length = 88
target-version = ['py311']
